win_acl_rs::sd

Struct SecurityDescriptorImpl

pub struct SecurityDescriptorImpl<P: PrivilegeLevel = Unprivileged> { /* private fields */ }

A Windows security descriptor.

Implementations

impl SecurityDescriptorImpl<Elevated>

pub fn from_path<P>(path: P) -> Result<Self, WinError>
where P: AsRef<Path>,

Creates a SecurityDescriptor from path to the “file object”

Arguments

path - Path to the file.

Returns

A SecurityDescriptor on success.

pub fn from_handle<S>( handle: S, object_type: SE_OBJECT_TYPE, ) -> Result<Self, WinError>
where S: AsRef<str>,

Creates a SecurityDescriptor from object name and object type.

Arguments

handle - name of the object. This could be many things (path to the file or directory, to network share, name of the printer, registry key, …)

object_type - a type of the object

see MSDN

Returns

A SecurityDescriptor on success.

impl SecurityDescriptorImpl<Unprivileged>

pub fn upgrade( self, _token: &PrivilegeTokenImpl<Elevated>, ) -> SecurityDescriptorImpl<Elevated>

Upgrades security descriptor to elevated one that has access to SACL, for example

Is guarded by the existence of elevated token

pub fn from_path<P>(path: P) -> Result<Self, WinError>
where P: AsRef<Path>,

Creates a SecurityDescriptor from path to the “file object”

Arguments

path - Path to the file.

Returns

A SecurityDescriptor on success.

pub fn from_handle<S>( handle: S, object_type: SE_OBJECT_TYPE, ) -> Result<Self, WinError>
where S: AsRef<str>,

Creates a SecurityDescriptor from object name and object type.

Arguments

handle - name of the object. This could be many things (path to the file or directory, to network share, name of the printer, registry key, …)

object_type - a type of the object

see MSDN

Returns

A SecurityDescriptor on success.

impl<P: PrivilegeLevel> SecurityDescriptorImpl<P>

pub fn is_valid(&self) -> bool

Validates a security descriptor.

pub fn owner_defaulted(&self) -> Result<bool, WinError>

Indicates that the SID of the owner of the security descriptor was provided by a default mechanism.

pub fn group_defaulted(&self) -> Result<bool, WinError>

Indicates that the security identifier (SID) of the security descriptor group was provided by a default mechanism.

pub fn dacl_defaulted(&self) -> Result<bool, WinError>

Indicates a security descriptor with a default DACL.

pub fn dacl_present(&self) -> Result<bool, WinError>

Indicates a security descriptor that has a DACL. If this flag is not set, or if this flag is set and the DACL is NULL, the security descriptor allows full access to everyone.

pub fn sacl_defaulted(&self) -> Result<bool, WinError>

Indicates a security descriptor with a default SACL.

pub fn sacl_present(&self) -> Result<bool, WinError>

Indicates a security descriptor that has a SACL.

pub fn from_sd_string<S>(sd_string: S) -> Result<Self, WinError>
where S: AsRef<str>,

Converts a string-format security descriptor into a valid, functional security descriptor.

see MSDN

Returns

A SecurityDescriptor on success.

pub fn as_sd_string(&self) -> Result<String, WinError>

Converts security descriptor into a string format

see MSDN

Returns

A String on success.

pub fn owner_sid(&self) -> Option<&Sid>

TODO

pub fn group_sid(&self) -> Option<&Sid>

TODO

pub fn dacl(&self) -> Option<&Acl>

TODO