win_acl_rs

Module sid

Security Identifier (SID) operations.

A SID is a unique value of variable length used to identify a security principal or group in Windows security contexts. This module provides safe wrappers for creating, parsing, and manipulating SIDs, as well as looking up account names from SIDs and vice versa.

Examples

use win_acl_rs::sid::Sid;
use win_acl_rs::wellknown::WinWorldSid;

// Create a SID from a well-known SID constant
let sid = Sid::from_well_known_sid(WinWorldSid)?; // Everyone SID

// Or create from a string representation
let admin_sid = Sid::from_string("S-1-5-32-544")?; // Administrators

// Look up the account name
let lookup = sid.lookup_name()?;
println!("Account: {}/{}", lookup.domain, lookup.name);

// Convert to string
let sid_string = sid.to_string()?;
println!("SID: {}", sid_string);

Modules

account

Structs

Sid

An owned Security Identifier (SID).

SidRef

A borrowed reference to a Security Identifier (SID).

Traits

AsSidRef

Trait for types that can be converted to a SidRef.