Crate win_acl_rs

This crate provides a simple and safe abstraction over Windows ACLs and security descriptors.

See examples folder

Modules

acl

elevated

this module exposes functions that require elevated privileges the process requires “SE_SECURITY_NAME” (SeSecurityPrivilege) privilege, would otherwise return WIN32_ERROR(1314) => “A required privilege is not held by the client” you can run whoami /priv to check it. You typically need to run the process as an Administrator and enable it using enable_se_security_privilege().

error

sd

sid

utils

Constants

SE_DS_OBJECT

SE_DS_OBJECT_ALL

SE_FILE_OBJECT

SE_KERNEL_OBJECT

SE_LMSHARE

SE_PRINTER

SE_PROVIDER_DEFINED_OBJECT

SE_REGISTRY_KEY

SE_REGISTRY_WOW64_32KEY

SE_REGISTRY_WOW64_64KEY

SE_SERVICE

SE_UNKNOWN_OBJECT_TYPE

SE_WINDOW_OBJECT

Type Aliases

SE_OBJECT_TYPE